+++ docs/conf.py
@@ -15,7 +15,9 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('..'))
-autodoc_mock_imports = ['pandas','numpy','scipy','annoy','packaging','scipy.spatial','scipy.sparse','sklearn','sklearn.neighbors','sklearn.linear_model','umap','umap.umap_']
+autodoc_mock_imports = ['pandas','numpy','scipy','annoy','packaging','scipy.spatial',
+						'scipy.sparse','sklearn','sklearn.neighbors','sklearn.linear_model',
+						'umap','umap.umap_','anndata','scanpy','pynndescent']
 
 # -- Project information -----------------------------------------------------
 
@@ -26,7 +28,7 @@ author = 'Krzysztof Polanski, Jongeun Park'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '1.4.1'
+release = '1.5.0'
 
 
 # -- General configuration ---------------------------------------------------